executeQuery slightly better
